Add change todo dialog widget
